refactor: move parts API to the epoch manager
cc https://github.com/near/nearcore/issues/6910
